requirements.txt
hydra-core==1.3.2
hydra-slayer==0.4.1
loguru==0.5.3
autogluon